Used my new `all_choices` macro to auto-gen combinatorial impl explosion.
The Rust macro system has some rough corners.  I am not thrilled with
the treatment of comma (and/or lack of expressiveness regarding when
commas are required and are optional).  I *think* that an oddity
related to this is what led me to adopt the hack of always putting an
ignored base case on my constructed list of identifiers: while I could
go from `(z)` to `(y,z)` and `(x,y,z)`, I could not go from `()` to
`(z,)` nor `(z)`.

----

Also did some other macro-hacking to ease the generation of the `row`
and `col` methods for the various dimension-specialized matrix
implementations.  However some weaknesses in the macro system led me
to take a side-track hack and break my desired `Matrix` trait up into
two traits: `Columns` and `Rows`, basically to divide and conquer the
effort of generating method code for the `col` and `row` accessors
which tend to be very uniform in one of the two dimensions (i.e. "row"
or "col") and somewhat non-uniform in the other dimension.  It may be
feasible to actually auto-generate the whole method impl by building
up the fields for the struct construction expression incrementally,
but since I only have to deal with the three cases of {2, 3, and 4},
hard-coding the implementations (apart from parameterizing over the
magnitude of the opposite dimension) seemed easier.

I did have to resort to using the type `tt` for the input to the
driver macro `impl_ColRow_for`, since apparently the macro system does
not like to first type something as `$expr` and then later try to pass
it into a macro expecting a more-specific input like `2`.  An
interesting detail and I think a sane one, but something to be aware
of when developing macros.